add FixedRewardPool

+++ contracts/test/TestDeviceNFT.sol
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
 import {ERC721} from "@openzeppelin/contracts/token/ERC721/ERC721.sol";
@@ -27,7 +28,7 @@ contract TestDeviceNFT is IWeightedNFT, ERC721 {
     weightOf[_tokenId] = _weight;
   }
 
-  function mint(address to, uint tokenId) external {
+  function mint(address to, uint256 tokenId) external {
     _mint(to, tokenId);
     weightOf[tokenId] = 1 ether;
   }
